% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bootstrap_estimator}
\alias{bootstrap_estimator}
\title{Bootstrap parameter estimates for confidence intervals.}
\usage{
bootstrap_estimator(rolypoly, bootstrap_iters, run_light, run_parallel)
}
\arguments{
\item{rolypoly}{rolypoly object}

\item{bootstrap_iters}{number of bootstrap iterations to run}

\item{run_light}{if we throw away bootstrap data}

\item{run_parallel}{if we want to collect bootstrap data in parallel}
}
\description{
Bootstrap parameter estimates for confidence intervals.
}
\examples{
\dontrun{bootstrap_estimator(rolypoly)}
}
