% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calculate_annotation_block_heritability}
\alias{calculate_annotation_block_heritability}
\title{Caclulate the contribution of block annotations to the heritability of a trait.}
\usage{
calculate_annotation_block_heritability(block_data, params)
}
\arguments{
\item{block_data}{functional information of blocks}

\item{params}{parameter fit}
}
\description{
A vector of indenpendent heritability contributions of block annotations is returned. Sum
the vector to get total explained heritability and divide by sum to get proportion.
}
\examples{
\dontrun{calculate_annotation_block_heritability(block_data, params)}
}
