% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calculate_block_values}
\alias{calculate_block_values}
\title{Caclulate predicted block values based on block information and model fit.}
\usage{
calculate_block_values(block_data, params)
}
\arguments{
\item{block_data}{block_data}

\item{params}{parameter fit}
}
\value{
returns block values
}
\description{
From a model fit we can predict expected variance of a block based on information
we have about the block. In the example of gene expression this would equate
to predicting the importance of a gene based on its signature of expression.
}
\examples{
\dontrun{calculate_gene_values(block_data, params)}
}
