% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calculate_expected_block_values_given_ld}
\alias{calculate_expected_block_values_given_ld}
\title{Caclulate predicted block values based on block information and model fit.}
\usage{
calculate_expected_block_values_given_ld(rolypoly, block_values)
}
\arguments{
\item{rolypoly}{rolypoly object}

\item{block_values}{estimated block values.}
}
\description{
From a model fit we can predict expected variance of a block based on information
we have about the block. In the example of gene expression this would equate
to predicting the importance of a gene based on its signature of expression.
}
\examples{
\dontrun{calculate_expected_block_values_given_ld(rolypoly, block_values)}
}
