% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularized_inference.R
\name{cv_regularized_parameter_estimator}
\alias{cv_regularized_parameter_estimator}
\title{Perform regularization inference.}
\usage{
cv_regularized_parameter_estimator(vectorized_rolypoly_data, n_folds = 10,
  ...)
}
\arguments{
\item{vectorized_rolypoly_data}{rolypoly data used for inference}

\item{n_folds}{number of folds for cross validation}

\item{...}{other arguments to pass to cv.glmnet}
}
\value{
results from cross validation
}
\description{
Use CV to find appropriate values of lambda for either feature selection
or for prediction.
}
\examples{
\dontrun{cv_regularized_parameter_estimator(vectorized_rolypoly_data)}
}
