% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{data_io}
\alias{data_io}
\title{Functions for opening and organizing data.}
\description{
We must open block annotation data, gwas data, snp annotations, gene annotations.
Here, you'll find functions that to this and organize these data into a rolypoly object.
}
