% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rolypoly_version.R
\name{is_rolypoly_pascaltimshel_fork}
\alias{is_rolypoly_pascaltimshel_fork}
\title{Find out what installation of rolypoly you have loaded}
\usage{
is_rolypoly_pascaltimshel_fork()
}
\arguments{
\item{dummy}{this function takes parameters}
}
\value{
prints statement that you are using pascaltimshel/rolypoly fork (and not dcalderon/rolypoly)
}
\description{
Some description...
}
\details{
This function enables to you check if the pascaltimshel rolypoly fork is loaded.
The following command will return true only if pascaltimshel rolypoly fork is loaded.
\code{exists('is_rolypoly_pascaltimshel_fork', mode='function')}
}
\examples{
\dontrun{dont run this ....}
}
