% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{make_ld_matrix}
\alias{make_ld_matrix}
\title{Helper function to pull LD data from NCBI.}
\usage{
make_ld_matrix(all_snps, ld_data)
}
\arguments{
\item{all_snps}{The snps that were queried}

\item{ld_data}{A returned LD matrix with SNP, Proxy, and RSquared columns}
}
\value{
an LD matrix where query snps will be the first columns in the correct order
}
\description{
Given the path of a gwas file open it into a data.table object
}
\examples{
\dontrun{make_ld_matrix(all_snps, ld_data)}
}
