% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{make_results_df}
\alias{make_results_df}
\title{Helper function to make a summary table of results from bootstrap data.}
\usage{
make_results_df(value_collection, annotations, model_estimates)
}
\arguments{
\item{value_collection}{collection of bootstrapped value estimates}

\item{annotations}{vector of annotation names}

\item{model_estimates}{estimates for bias parameter estimates}
}
\value{
data frame with results summary
}
\description{
Helper function to make a summary table of results from bootstrap data.
}
\examples{
\dontrun{make_results_df(value_collection)}
}
