% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{parameter_estimator}
\alias{parameter_estimator}
\title{Find parameter estimates for the data.}
\usage{
parameter_estimator(vectorized_rolypoly_data)
}
\arguments{
\item{vectorized_rolypoly_data}{rolypoly data that has been vectorized}
}
\value{
results of inference
}
\description{
Find parameter estimates for the data.
}
\examples{
\dontrun{parameter_estimator(rolypoly)}
}
