% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolypoly_plots.R
\name{plot_rolypoly_annotation_estimates}
\alias{plot_rolypoly_annotation_estimates}
\title{Visualize parameter estimates after running inference}
\usage{
plot_rolypoly_annotation_estimates(rolypoly)
}
\arguments{
\item{rolypoly}{a rolypoly object}
}
\value{
ggplot2 object
}
\description{
Visualize parameter estimates after running inference
}
\examples{
\dontrun{plot_rolypoly_annotation_estimates(rolypoly)}
}
