% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolypoly_plots.R
\name{plot_rolypoly_annotation_ranking}
\alias{plot_rolypoly_annotation_ranking}
\title{Rank annotations by p-value after running inference}
\usage{
plot_rolypoly_annotation_ranking(rolypoly)
}
\arguments{
\item{rolypoly}{a rolypoly object}
}
\value{
ggplot2 object
}
\description{
Rank annotations by p-value after running inference
}
\examples{
\dontrun{plot_rolypoly_annotation_ranking(rolypoly)}
}
