% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularized_inference.R
\name{regularized_inference}
\alias{regularized_inference}
\title{Inference functions that include regularization}
\description{
Functions for inferring relevant annotations using the polyTest model.
}
