% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{robust_parameter_estimator}
\alias{robust_parameter_estimator}
\title{Find robust parameter estimates for the data.}
\usage{
robust_parameter_estimator(vectorized_rolypoly_data, outlier_threshold = 10)
}
\arguments{
\item{vectorized_rolypoly_data}{vectorized rolypoly data}

\item{outlier_threshold}{outlier threshold for robust inference}
}
\description{
Find robust parameter estimates for the data.
}
\examples{
\dontrun{parameter_estimator(rolypoly)}
}
