% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolypoly.R
\docType{package}
\name{rolypoly}
\alias{rolypoly}
\alias{rolypoly-package}
\title{The rolypoly package for identifying annotations associated with complex traits.}
\description{
The rolypoly package for identifying annotations associated with complex traits.
}
