% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{rolypoly_add_ld_corrected_gwas_block_scores}
\alias{rolypoly_add_ld_corrected_gwas_block_scores}
\title{Add LD corrected block scores to rolypoly.}
\usage{
rolypoly_add_ld_corrected_gwas_block_scores(rolypoly, fast_calculation = T)
}
\arguments{
\item{rolypoly}{rolypoly data}

\item{fast_calculation}{if F then LD deconvolution else quadratic form.}
}
\value{
rolypoly object with LD corrected gwas scores attached
}
\description{
Add LD corrected block scores to rolypoly.
}
\examples{
\dontrun{rolypoly_add_ld_corrected_gwas_block_scores(rolypoly)}
}
