% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{rolypoly_link_blocks_and_gwas}
\alias{rolypoly_link_blocks_and_gwas}
\title{Link blocks and gwas}
\usage{
rolypoly_link_blocks_and_gwas(rolypoly, ld_folder, r2_threshold = 0.2,
  run_parallel = F)
}
\arguments{
\item{rolypoly}{a rolypoly object}

\item{ld_folder}{path to a folder with ld data}

\item{r2_threshold}{LD threshold to look at data}

\item{run_parallel}{check if user wants to run in parallel}
}
\value{
rolypoly object with data attached
}
\description{
Takes block information, potentially independent LD blocks or gene blocks,
and gwas data and organizes the data for interenal processing
}
\examples{
\dontrun{rolypoly_link_blocks_and_gwas(rolypoly, ld_folder, r2_threshold)}
}
