% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{rolypoly_load_block_annotation}
\alias{rolypoly_load_block_annotation}
\title{Load annotations for blocks of LD, in some cases this is a gene annotation
with a window around a gene.}
\usage{
rolypoly_load_block_annotation(rolypoly, block_annotation, genes = T)
}
\arguments{
\item{rolypoly}{rolypoly data object}

\item{block_annotation}{annotation information for block}

\item{genes}{if these are genes}
}
\value{
rolypoly data with block annotations attached
}
\description{
Load annotations for blocks of LD, in some cases this is a gene annotation
with a window around a gene.
}
\examples{
\dontrun{rolypoly_load_block_annotation(rolypoly, block_annotation)}
}
