% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{rolypoly_load_gwas}
\alias{rolypoly_load_gwas}
\title{Load gwas data}
\usage{
rolypoly_load_gwas(rolypoly, gwas_data, snp_annotations = NULL,
  gwas_z_filter = -1, add_spline = F, n_knots = 1, add_poly = F,
  n_degree = 2)
}
\arguments{
\item{rolypoly}{rolypoly data}

\item{gwas_data}{gwas data}

\item{snp_annotations}{if there are additional snp annotations included}

\item{gwas_z_filter}{if we want to remove large effect SNPs}

\item{add_spline}{for fitting a spline to maf}

\item{n_knots}{number of knots for spline}

\item{add_poly}{for fitting a polynomial to maf}

\item{n_degree}{degree of polynomial to fit}
}
\value{
rolypoly object with gwas data loaded
}
\description{
Load gwas data
}
\examples{
\dontrun{rolypoly_load_gwas(rolypoly, gwas_data)}
}
