% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{rolypoly_perform_inference}
\alias{rolypoly_perform_inference}
\title{Run inference.}
\usage{
rolypoly_perform_inference(rolypoly, bootstrap_iters = 50,
  outlier_threshold = -1, run_light = F, run_parallel = F)
}
\arguments{
\item{rolypoly}{rolypoly object}

\item{bootstrap_iters}{number of bootstrap iterations to perform}

\item{outlier_threshold}{threshold for performing robust regression, still experimental.}

\item{run_light}{if we throw out bootstrap data}

\item{run_parallel}{if we collect bootstraps in parallel}
}
\value{
rolypoly object with inference information attached
}
\description{
Coordinates running inference.
}
\examples{
\dontrun{run_inference(rolypoly)}
}
