% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularized_inference.R
\name{rolypoly_perform_regularized_inference}
\alias{rolypoly_perform_regularized_inference}
\title{Run inference with added regularization.}
\usage{
rolypoly_perform_regularized_inference(rolypoly, ...)
}
\arguments{
\item{rolypoly}{a rolypoly object}

\item{...}{other arguments to pass to cv.glmnet}
}
\value{
rolypoly object with regularization results
}
\description{
If p-values are desired use the other inference function. This for
prediction purposes.
}
\examples{
\dontrun{rolypoly_perform_regularized_inference(rolypoly)}
}
