% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_block_annotation.R
\docType{data}
\name{sim_block_annotation}
\alias{sim_block_annotation}
\title{Simulated block data annotation.}
\format{A data frame with 1000 rows and 4 variables:
\describe{
  \item{chrom}{chromosome, we only use autosomes}
  \item{start}{base pair position of variant}
  \item{end}{rsid identifier of variant}
  \item{label}{effect size, univariate regression coefficient}
}}
\source{
I generated these fields to link with SNP positions
}
\usage{
sim_block_annotation
}
\description{
A dataset containing simulated block data annotation for
 use in the rolypoly vignette.
}
\keyword{datasets}
