% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_expression_data.R
\docType{data}
\name{sim_expression_data_normalized}
\alias{sim_expression_data_normalized}
\title{Simulated expression data.}
\format{A data frame with 1000 rows and 5 variables:
\describe{
  \item{Liver}{simulated expression data for tissue. This tissue has a significant effect on the simulated gwas.}
  \item{Adrenal.Gland}{simulated expression data for tissue}
  \item{Blood}{simulated expression data for tissue. This tissue has a significant effect on the simulated gwas.}
  \item{Heart}{simulated expression data for tissue}
  \item{Lung}{simulated expression data for tissue}
}}
\source{
I generated this dataset
}
\usage{
sim_expression_data_normalized
}
\description{
A dataset containing simulated expression data for
 use in the rolypoly vignette.
}
\keyword{datasets}
