% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gwas_data.R
\docType{data}
\name{sim_gwas_data}
\alias{sim_gwas_data}
\title{Simulated GWAS summary statistics}
\format{A data frame with 14934 rows and 6 variables:
\describe{
  \item{chrom}{chromosome, we only use autosomes}
  \item{pos}{base pair position of variant}
  \item{rsid}{rsid identifier of variant}
  \item{beta}{effect size, univariate regression coefficient}
  \item{se}{standard error of effect size}
  \item{maf}{minor allele frequency}
}}
\source{
rsids were from 1000g and I generated the other fields
}
\usage{
sim_gwas_data
}
\description{
A dataset containing simulated genome-wide association summary statistics for
 use in the rolypoly vignette.
}
\keyword{datasets}
