% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{vectorize_rolypoly}
\alias{vectorize_rolypoly}
\title{Take a list of rolypoly data and vectorize it for inference.}
\usage{
vectorize_rolypoly(data)
}
\arguments{
\item{data}{the list of block information from rolypoly object}
}
\value{
list of necessary information for inference
}
\description{
Take a list of rolypoly data and vectorize it for inference.
}
\examples{
\dontrun{vectorize_rolypoly(data)}
}
